requirements.txt
openai>=1.0.0
python-dotenv~=1.1.0
pytz
requests~=2.32.3
